Change method Token#transfer to Wallet#transfer
