+++ test/wallet.spec.ts
@@ -453,4 +453,93 @@ describe('Wallet', () => {
       });
     });
   });
+
+  describe('transfer', () => {
+    const colorId1 =
+      'c1ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff';
+    const colorId2 =
+      'c2ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff';
+    const utxos = [
+      new Utxo(
+        '2222222222222222222222222222222222222222222222222222222222222222',
+        102,
+        1,
+        '76a914e18c333242b4d4ecbdc9b7071743b5bce53ea0ad88ac',
+        '000000000000000000000000000000000000000000000000000000000000000000',
+        10_000,
+      ),
+      new Utxo(
+        '9b9ebcb5bb47bf78b94fd696186926abc1daf756d833e90268d97fc1b370eca7',
+        0,
+        0,
+        '21c2ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffbc76a914e18c333242b4d4ecbdc9b7071743b5bce53ea0ad88ac',
+        'c2ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff',
+        1_000,
+      ),
+      new Utxo(
+        '9b9ebcb5bb47bf78b94fd696186926abc1daf756d833e90268d97fc1b370eca7',
+        0,
+        1,
+        '21c1ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffbc76a914e18c333242b4d4ecbdc9b7071743b5bce53ea0ad88ac',
+        'c1ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff',
+        3_000,
+      ),
+    ];
+    const wif = 'KzJYKvdPEkuDanYNecre9QHe4ugRjvMvoLeceRr4j5u2j9gEyQ7n';
+    const toAddress = '12URZZB1vXWTdqFiQqJCNPHftX2uAnWMos';
+    const amount = 1;
+    const changePubkeyScript = Buffer.from(
+      '76a914e18c333242b4d4ecbdc9b7071743b5bce53ea0ad88ac',
+      'hex',
+    );
+
+    describe('transfer single token', () => {
+      it('should build transaction', async () => {
+        const { wallet: alice, dataStore } = createWallet('prod');
+        await alice.importWif(wif);
+        await dataStore.add(utxos);
+        const result = await alice.transfer(
+          [
+            {
+              colorId: colorId1,
+              amount: amount,
+              toAddress: toAddress,
+            },
+          ],
+          changePubkeyScript,
+        );
+        const tx = result.txb.buildIncomplete();
+        console.log(tx);
+        assert.strictEqual(tx.ins.length, 2);
+        assert.strictEqual(tx.outs.length, 3);
+      });
+    });
+
+    describe('transfer multiple token', () => {
+      it('should build transaction', async () => {
+        const { wallet: alice, dataStore } = createWallet('prod');
+        await alice.importWif(wif);
+        await dataStore.add(utxos);
+        const result = await alice.transfer(
+          [
+            {
+              colorId: colorId1,
+              amount: amount,
+              toAddress: toAddress,
+            },
+            {
+              colorId: colorId2,
+              amount: amount,
+              toAddress: toAddress,
+            },
+          ],
+          changePubkeyScript,
+        );
+        const tx = result.txb.buildIncomplete();
+        console.log(tx);
+        assert.strictEqual(tx.ins.length, 3);
+        assert.strictEqual(tx.outs.length, 5);
+      });
+    });
+  });
 });
