+++ ts_src/index.ts
@@ -2,7 +2,6 @@ import { Config } from './config';
 import * as DataStore from './data_store';
 import * as KeyStore from './key_store';
 import * as Signer from './signer';
-import { BaseToken, Token } from './token';
 import * as Wallet from './wallet';
 
-export { DataStore, KeyStore, Wallet, Config, Signer, BaseToken, Token };
+export { DataStore, KeyStore, Wallet, Config, Signer };
