+++ types/token.d.ts
@@ -1,14 +1,5 @@
-/// <reference types="node" />
 import * as tapyrus from 'tapyrusjs-lib';
-import { Utxo } from './utxo';
-import Wallet from './wallet';
-export interface Token {
-    transfer(wallet: Wallet, toAddress: string, amount: number, changePubkeyScript: Buffer): Promise<{
-        txb: tapyrus.TransactionBuilder;
-        inputs: Utxo[];
-    }>;
-}
-export declare class TokenParams {
+export declare class TransferParams {
     colorId: string;
     amount: number;
     toAddress: string;
@@ -19,12 +10,4 @@ export declare const TokenTypes: {
     NON_REISSUABLE: number;
     NFT: number;
 };
-export declare class BaseToken {
-    transfer(wallet: Wallet, params: TokenParams[], changePubkeyScript: Buffer): Promise<{
-        txb: tapyrus.TransactionBuilder;
-        inputs: Utxo[];
-    }>;
-    private addressToOutput;
-    private collect;
-    private createDummyTransaction;
-}
+export declare function createDummyTransaction(txb: tapyrus.TransactionBuilder): tapyrus.Transaction;
