+++ types/wallet.d.ts
@@ -1,9 +1,11 @@
+/// <reference types="node" />
 import * as tapyrus from 'tapyrusjs-lib';
 import { Balance } from './balance';
 import { Config } from './config';
 import { DataStore } from './data_store';
 import { KeyStore } from './key_store';
 import { Rpc } from './rpc';
+import { TransferParams } from './token';
 import { Utxo } from './utxo';
 export default interface Wallet {
     keyStore: KeyStore;
@@ -15,6 +17,10 @@ export default interface Wallet {
     balance(colorId?: string): Promise<Balance>;
     utxos(colorId?: string): Promise<Utxo[]>;
     estimatedFee(tx: tapyrus.Transaction): number;
+    transfer(params: TransferParams[], changePubkeyScript: Buffer): Promise<{
+        txb: tapyrus.TransactionBuilder;
+        inputs: Utxo[];
+    }>;
 }
 export declare class BaseWallet implements Wallet {
     static COLOR_ID_FOR_TPC: string;
@@ -29,7 +35,13 @@ export declare class BaseWallet implements Wallet {
     broadcast(tx: tapyrus.Transaction): Promise<string>;
     balance(colorId?: string): Promise<Balance>;
     utxos(colorId?: string): Promise<Utxo[]>;
+    transfer(params: TransferParams[], changePubkeyScript: Buffer): Promise<{
+        txb: tapyrus.TransactionBuilder;
+        inputs: Utxo[];
+    }>;
     estimatedFee(tx: tapyrus.Transaction): number;
     private listUnspent;
     private privateToScriptHash;
+    private addressToOutput;
+    private collect;
 }
